@@ -20,9 +20,9 @@
 	"devDependencies": {
 		"@sveltejs/adapter-auto": "^3.0.0",
 		"@sveltejs/adapter-node": "^2.1.0",
-		"@sveltejs/enhanced-img": "^0.2.0",
-		"@sveltejs/kit": "^2.0.0",
-		"@sveltejs/vite-plugin-svelte": "^3.0.0",
+		"@sveltejs/enhanced-img": "^0.3.9",
+		"@sveltejs/kit": "^2.7.2",
+		"@sveltejs/vite-plugin-svelte": "^4.0.0",
 		"@types/bootstrap": "^5.2.10",
 		"@types/eslint": "^8.56.0",
 		"@typescript-eslint/eslint-plugin": "^7.0.0",
@@ -30,21 +30,22 @@
 		"bootstrap": "^5.3.2",
 		"eslint": "^8.56.0",
 		"eslint-config-prettier": "^9.1.0",
-		"eslint-plugin-svelte": "^2.35.1",
+		"eslint-plugin-svelte": "^2.45.1",
 		"prettier": "^3.1.1",
-		"prettier-plugin-svelte": "^3.1.2",
+		"prettier-plugin-svelte": "^3.2.6",
 		"sass": "^1.77.1",
-		"svelte": "^4.2.7",
-		"svelte-check": "^3.6.0",
+		"svelte": "^5.0.0",
+		"svelte-check": "^4.0.0",
 		"tslib": "^2.4.1",
-		"typescript": "^5.0.0",
-		"vite": "^5.0.3"
+		"typescript": "^5.5.0",
+		"vite": "^5.4.4"
 	},
 	"type": "module",
 	"dependencies": {
 		"@tabler/core": "^1.0.0-beta20",
-		"@tabler/icons-svelte": "^3.5.0",
+		"@tabler/icons-svelte": "^3.20.0",
+		"@tanstack/svelte-query": "^5.59.16",
 		"svelte-bootstrap-icons": "^3.1.1",
-		"svelte-preprocess": "^5.1.4"
+		"svelte-preprocess": "^6.0.0"
 	}
 }
